cffi/requirements.txt
cffi
attrs
jinja2